Move MBM conversion into a `PsiLuaEnv` extension
